ConnectionHelper moved into utils package; Database name is now the first arg in getMongoTemplate
